~ extract Modbus value access
Move process-image value conversion and IndexRange handling out of
ModbusAddressSpace so the address-space fragment stays focused on OPC UA
dispatch, lifecycle, and persistence.
